requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.1.1
python-dotenv>=0.19.0
gunicorn>=20.1.0
requests>=2.31.0